Remove get_current_context
Unfortunately this is incompatible with the impl of Drop on Window. Creating a new Window with the same handle would cause the TLS to be cleared from under us. If you'd like to know whether a context is current, you can use the Window::is_current method implemented in the last commit.